requirements.txt
alabaster==0.7.10
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
coveralls==1.2.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
django-material==1.1.1
django-registration-redux==1.8
django-toolbelt==0.0.1
docopt==0.6.2
docutils==0.12
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
imagesize==0.7.1
invoke==0.12.2
isort==4.2.15
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
pkginfo==1.4.1
pluggy==0.5.2
psycopg2==2.7.3.1
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.2.3
pytz==2017.2
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
static3==0.7.0
tox==2.9.1
tox-travis==0.8
tqdm==4.19.4
twine==1.9.1
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.1
